CLN: consolidate arrow roundtrip tests for nullable dtypes in base masked tests
